Task2/requirements.txt
# PDF Processing
PyPDF2

# Google Gemini API
google-generativeai

# Scientific Computing and ML
numpy
scikit-learn

# Environment variables (optional)
python-dotenv
